both binarySearch

+++ src/GenTree.cpp
@@ -1,11 +1,11 @@
 #include "GenTree.h"
 
 byte GenTree::bit_count[256];
-int *GenTree::roots;
-int *GenTree::left;
-int *GenTree::ryte;
-int *GenTree::parent;
-int GenTree::ixRoots;
-int GenTree::ixLeft;
-int GenTree::ixRyte;
-int GenTree::ixPrnt;
+int16_t *GenTree::roots;
+int16_t *GenTree::left;
+int16_t *GenTree::ryte;
+int16_t *GenTree::parent;
+int16_t GenTree::ixRoots;
+int16_t GenTree::ixLeft;
+int16_t GenTree::ixRyte;
+int16_t GenTree::ixPrnt;
